+++ lib/ui/page/home/page/user/controller.dart
@@ -32,7 +32,6 @@ import '/domain/model/link.dart';
 import '/domain/model/mute_duration.dart';
 import '/domain/model/my_user.dart';
 import '/domain/model/precise_date_time/precise_date_time.dart';
-import '/domain/model/price.dart';
 import '/domain/model/user.dart';
 import '/domain/repository/contact.dart';
 import '/domain/repository/paginated.dart';
@@ -138,6 +137,9 @@ class UserController extends GetxController {
   /// [Paginated] containing the [DirectLink] leading to this [MyUser].
   late final Paginated<DirectLinkSlug, DirectLink> links;
 
+  /// Indicator whether a [Block] describing promo program should be expanded.
+  final RxBool promoExpanded = RxBool(true);
+
   /// [UserService] fetching the [user].
   final UserService _userService;
 
@@ -196,6 +198,11 @@ class UserController extends GetxController {
   /// Returns the individual [MonetizationSettings] for separate [UserId]s.
   RxMap<UserId, Rx<MonetizationSettings>> get individual =>
       _partnerService.individual;
+
+  /// Returns the [MonetizationSettings] that the [UserId]s have for our
+  /// [MyUser].
+  RxMap<UserId, Rx<MonetizationSettings>> get monetization =>
+      _partnerService.monetization;
 
   @override
   void onInit() {
@@ -431,19 +438,15 @@ class UserController extends GetxController {
 
   /// Updates the [MonetizationSettings] for the current [User].
   Future<void> updateMonetizationSettings({
-    bool? donationsEnabled,
-    Sum? donationsMinimum,
+    MonetizationSettingsDonation? donation,
+    MonetizationSettingsMessage? message,
+    MonetizationSettingsReferral? referral,
   }) async {
     await _partnerService.updateMonetizationSettings(
       userId: id,
-      donation: NewType(
-        donationsEnabled == null && donationsMinimum == null
-            ? null
-            : MonetizationSettingsDonation(
-                enabled: donationsEnabled ?? true,
-                min: Price.xxx(donationsMinimum?.val ?? 1),
-              ),
-      ),
+      donation: NewType(donation),
+      message: NewType(message),
+      referral: NewType(referral),
     );
   }